+++ src/ResourceManager/RecoveryServices.SiteRecovery/Commands.RecoveryServices.SiteRecovery/Job/GetAzureRmRecoveryServicesAsrJob.cs
@@ -33,7 +33,7 @@ namespace Microsoft.Azure.Commands.RecoveryServices.SiteRecovery
     public class GetAzureRmRecoveryServicesAsrJob : SiteRecoveryCmdletBase
     {
         /// <summary>
-        ///     Gets or sets Job Name.
+        ///     Gets the details of the specified ASR job or the list of recent ASR jobs in the Recovery Services vault.
         /// </summary>
         [Parameter(
             ParameterSetName = ASRParameterSets.ByName,
@@ -52,8 +52,7 @@ namespace Microsoft.Azure.Commands.RecoveryServices.SiteRecovery
         public ASRJob Job { get; set; }
 
         /// <summary>
-        ///     Gets or sets start time. Allows to filter the list of jobs started after the given
-        ///     start time.
+        ///     Gets or sets start time. Allows to filter the list of jobs started after the given start time.
         /// </summary>
         [Parameter(
             ParameterSetName = ASRParameterSets.ByParam,