+++ src/ResourceManager/DevTestLabs/ChangeLog.md
@@ -19,6 +19,9 @@
 -->
 ## Current Release
 
+## Version 4.0.3
+* Fix issue with Default Resource Group in CloudShell
+
 ## Version 4.0.2
 * Fixed all of the GET cmdlets which previously weren't working.